# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Avito PVZ Service 2025",
        "title": "PVZ Service",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/pvz": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список ПВЗ с их приемками и товарами с возможностью фильтрации по дате и пагинацией",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pvz"
                ],
                "summary": "Получить список ПВЗ с пагинацией",
                "operationId": "get-pvz-with-pagination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Начальная дата (RFC3339)",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Конечная дата (RFC3339)",
                        "name": "endDate",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Номер страницы (по умолчанию 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Количество элементов на странице (по умолчанию 10, максимум 30)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список ПВЗ с приемками и товарами",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PVZWithReceptionsDto"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новый пункт выдачи заказов в указанном городе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pvz"
                ],
                "summary": "Завести ПВЗ",
                "operationId": "create-pvz",
                "parameters": [
                    {
                        "description": "Информация о создаваемом ПВЗ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePVZRequestDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ПВЗ успешно создан",
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePVZResponseDto"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные / Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    }
                }
            }
        },
        "/pvz/{pvzId}/close_last_reception": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Закрывает последнюю активную приемку для указанного ПВЗ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pvz"
                ],
                "summary": "Закрыть последнюю приемку",
                "operationId": "close-last-reception",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор ПВЗ",
                        "name": "pvzId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Приемка успешно закрыта",
                        "schema": {
                            "$ref": "#/definitions/dto.CloseReceptionResponseDto"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные / ПВЗ не найден / Приемка уже закрыта",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    }
                }
            }
        },
        "/pvz/{pvzId}/delete_last_product": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет последний добавленный товар из последней активной приемки для указанного ПВЗ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pvz"
                ],
                "summary": "Удалить последний товар",
                "operationId": "delete-last-product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор ПВЗ",
                        "name": "pvzId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Товар успешно удален"
                    },
                    "400": {
                        "description": "Некорректные данные / ПВЗ не найден / Нет активной приемки / Нет товаров для удаления",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CloseReceptionResponseDto": {
            "description": "Информация о приемке при ее закрытии",
            "type": "object",
            "properties": {
                "dateTime": {
                    "description": "Дата и время приемки",
                    "type": "string"
                },
                "id": {
                    "description": "Идентификатор приемки",
                    "type": "string"
                },
                "pvzId": {
                    "description": "Идентификатор ПВЗ",
                    "type": "string"
                },
                "status": {
                    "description": "Статус приемки",
                    "type": "string"
                }
            }
        },
        "dto.CreatePVZRequestDto": {
            "description": "Информация о ПВЗ при его создании",
            "type": "object",
            "properties": {
                "city": {
                    "description": "Город",
                    "type": "string"
                }
            }
        },
        "dto.CreatePVZResponseDto": {
            "description": "Информация о ПВЗ при его создании",
            "type": "object",
            "properties": {
                "city": {
                    "description": "Город",
                    "type": "string"
                },
                "id": {
                    "description": "Идентификатор",
                    "type": "string"
                },
                "registrationDate": {
                    "description": "Дата регистрации",
                    "type": "string"
                }
            }
        },
...
        "dto.ErrorDto": {
            "description": "Информация об ошибке (DTO)",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Текст ошибки",
                    "type": "string"
                }
            }
        },
        "dto.PVZDto": {
            "description": "Информация о ПВЗ",
            "type": "object",
            "properties": {
                "city": {
                    "description": "Город",
                    "type": "string"
                },
                "id": {
                    "description": "Идентификатор",
                    "type": "string"
                },
                "registrationDate": {
                    "description": "Дата регистрации",
                    "type": "string"
                }
            }
        },
        "dto.PVZWithReceptionsDto": {
            "description": "Информация о ПВЗ и приемках, связанных с ним",
            "type": "object",
            "properties": {
                "pvz": {
                    "description": "Информация о ПВЗ",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.PVZDto"
                        }
                    ]
                },
                "receptions": {
                    "description": "Информация о всех приемках на ПВЗ",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ReceptionWithProductsDto"
                    }
                }
            }
        },
        "dto.ProductDto": {
            "description": "Информация о товаре",
            "type": "object",
            "properties": {
                "dateTime": {
                    "description": "Дата и время",
                    "type": "string"
                },
                "id": {
                    "description": "Идентификатор",
                    "type": "string"
                },
                "receptionId": {
                    "description": "Идентификатор приемки",
                    "type": "string"
                },
                "type": {
                    "description": "Тип товара",
                    "type": "string"
                }
            }
        },
        "dto.ReceptionDto": {
            "description": "Информация о приемке",
            "type": "object",
            "properties": {
                "dateTime": {
                    "description": "Дата и время создания приемки",
                    "type": "string"
                },
                "id": {
                    "description": "Идентификатор приемки",
                    "type": "string"
                },
                "pvzId": {
                    "description": "Идентификатор ПВЗ",
                    "type": "string"
                },
                "status": {
                    "description": "Статус приемки",
                    "type": "string"
                }
            }
        },
        "dto.ReceptionWithProductsDto": {
            "description": "Информация о приемке и товарах в ней",
            "type": "object",
            "properties": {
                "products": {
                    "description": "Информация о всех товарах в приемке",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ProductDto"
                    }
                },
                "reception": {
                    "description": "Информация о приемке",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.ReceptionDto"
                        }
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Authorization check",
            "type": "apiKey",
            "name": "auth-x",
            "in": "header"
        }
    }
}